CS-31 moved from pep8 to flake8
