+++ cloudpassage/utility.py
@@ -146,7 +146,8 @@ def parse_status(url, resp_code, resp_text):
         bad_statuses = {500: CloudPassageInternalError(resp_text, code=500),
                         400: CloudPassageValidation(resp_text, code=400),
                         401: CloudPassageAuthentication(resp_text, code=401),
-                        404: CloudPassageResourceExistence(resp_text, code=404, url=url),
+                        404: CloudPassageResourceExistence(resp_text, code=404,
+                                                           url=url),
                         403: CloudPassageAuthorization(resp_text, code=403),
                         422: CloudPassageValidation(resp_text, code=422)}
         if resp_code in bad_statuses:
