+++ tests/integration/test_integration_alert_profile.py
@@ -1,7 +1,6 @@
 import cloudpassage
 import json
 import os
-import pytest
 
 config_file_name = "portal.yaml.local"
 tests_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), "../"))
